Rebase and rename locale.ts to theme.ts
